requirements.txt
accelerate==0.29.1
bitsandbytes==0.43.0
gradio==4.26.0
huggingface-hub==0.22.2
numpy==1.24.3
pymupdf==1.24.0
python-docx==1.1.0
sentence_transformers==2.6.1
torch==2.2.2+cu121
transformers==4.39.3
